Journal Fragment draft upload
This is the first complete draft of the visual side of the journal fragment, which includes all the text input fields and a button to submit.

From here, I must create a backend for collecting the data and saving it to a file/timeline.

From there, I can go to the timeline tab, and then work on reading the entries, by retrieving the data from the saved data location, wherever that ends up being.