Project Lighthouse Update
Basic functionality for nearly everything seems to be working.

The only basic functionality still missing, is.. I believe, the custom dialog for viewing the details of a journal entry on the timeline screen.

Currently, the dialog is blank, but after that gets fixed up, I can work on the bells, whistles, and UI.