Added Bells and Whistles / Bugfixes
Beyond future improvements, of which there are many possible...  I am already making a list.  I've added some bells and whistles like background and icon images, make some text more readable on certain colors, etc.

The notifications currently only get pushed when the user switches the affirmation color, but I am still working on that.  There is currently code in the application for said purpose, but it's in its very early attempt stages, so it's definitely not working.  It's an attempt based off of a hodgepodge of other user questions on places like StackOverflow.